Move test into windows_test.go
